Non-interactive mode is detected automatically; batch files easier to read
The second commandline parameter "batch" is no longer needed because the script can figure out on its own whether it is run in a sufficiently interactive way and keyboard input or delays should be used. The embedded C# code is taken from my ConsoleHelper class.

The start batch files have the ps1 file name and config parts at the top so that they won't be lost at the right end of a very long line.